update from review and refactor
