+++ vllm/entrypoints/openai/serving_chat.py
@@ -237,25 +237,16 @@ class OpenAIServingChat(OpenAIServing):
                 log_tracing_disabled_warning()
 
             if isinstance(sampling_params, BeamSearchParams):
-                if isinstance(self.engine_client, AsyncLLMEngine):
-                    result_generator = self.engine_client.beam_search(
-                        engine_inputs['prompt_token_ids'],
-                        request_id,
-                        sampling_params,
-                    )
-                elif isinstance(self.engine_client, MQLLMEngineClient):
-                    result_generator = self.engine_client.beam_search(
-                        engine_inputs['prompt_token_ids'],
-                        request_id,
-                        sampling_params,
-                        lora_request,
-                    )
-                else:
-                    raise ValueError(
-                        "Beam search in the API server is only supported with"
-                        " AsyncLLMEngine and MQLLMEngineClient. please add "
-                        "`--disable-frontend-multiprocessing` to "
-                        "use beam search.")
+                assert isinstance(self.engine_client,
+                                    (AsyncLLMEngine,
+                                    MQLLMEngineClient)), \
+                    "Beam search is only supported with" \
+                    "AsyncLLMEngine and MQLLMEngineClient."
+                result_generator = self.engine_client.beam_search(
+                    engine_inputs['prompt_token_ids'],
+                    request_id,
+                    sampling_params,
+                )
             else:
                 result_generator = self.engine_client.generate(
                     engine_inputs,
